+++ index.js
@@ -1,4 +1,6 @@
-var fs = require('fs');
+var fs = require('fs'), path = require('path');
+
+var parentDir = path.dirname(module.parent.filename); // Directory of the module from which requireAll is called from
 
 var DEFAULT_EXCLUDE_DIR = /^\./;
 var DEFAULT_FILTER = /^([^\.].*)\.js(on)?$/;
@@ -14,31 +16,33 @@ module.exports = function requireAll(options) {
   var map = options.map || identity;
 
   function excludeDirectory(dirname) {
-    return !recursive ||
-      (excludeDirs && dirname.match(excludeDirs));
+    return !recursive || (excludeDirs && dirname.match(excludeDirs));
   }
+  
+  var dirPath = dirname;
+  if(!path.isAbsolute(dirname)) dirPath = parentDir + '/' + dirname;
+  
+  var files = fs.readdirSync(dirPath);
+  files.forEach(function(file) {
+    var filePath = dirPath + '/' + file;
 
-  var files = fs.readdirSync(dirname);
+    if(fs.statSync(filePath).isDirectory()) {
+      if(excludeDirectory(file)) return;
 
-  files.forEach(function (file) {
-    var filepath = dirname + '/' + file;
-    if (fs.statSync(filepath).isDirectory()) {
-
-      if (excludeDirectory(file)) return;
-
-      modules[map(file, filepath)] = requireAll({
-        dirname: filepath,
+      modules[map(file, filePath)] = requireAll({
+        dirname: filePath,
         filter: filter,
         excludeDirs: excludeDirs,
         map: map,
         resolve: resolve
       });
-
     } else {
       var match = file.match(filter);
-      if (!match) return;
+      if(!match) return;
 
-      modules[map(match[1], filepath)] = resolve(require(filepath));
+      var name = map(match[1], filePath);
+
+      modules[name] = resolve(require(filePath), name, filePath);
     }
   });