+++ Changes.md
@@ -4,6 +4,10 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## Unreleased
+* Adding __dirname before the directory not longer necessary
+* Added name and path to resolve function
+
 ## v2.0.0 (2015-10-17)
 
 * Add "recursive" argument to control directory recursion #21